fix: refactor to make psbt conversions easier with legacy api
