+++ packages/caravan-psbt/src/psbtv0/index.ts
@@ -1,1 +1,2 @@
-export { getUnsignedMultisigPsbtV0 } from "./utils";
+export * from "./psbt";
+export * from "./utils";
