+++ packages/caravan-psbt/src/psbtv0/utils.ts
@@ -1,135 +1,160 @@
+/**
+ * @file This file primarily contains utility functions migrated from the
+ * legacy psbt module in @caravan/bitcoin. With the new @caravan/psbt
+ * module, the goal is to make a more modular and legible API. But in order
+ * to make migrations easier from the old API, we need to provide conversion functions
+ * for converting the deeply nested objects in the legacy API.
+ */
 import {
-  ExtendedPublicKey,
-  MultisigAddressType,
+  generateBip32DerivationByIndex,
+  generateBraid,
+  multisigAddressType,
+  multisigBraidDetails,
+  multisigRedeemScript,
+  multisigWitnessScript,
   Network,
-  networkData,
   P2SH,
+  P2SH_P2WSH,
+  P2WSH,
 } from "@caravan/bitcoin";
-import { Psbt, Transaction } from "bitcoinjs-lib";
-import { MultisigWalletConfig } from "@caravan/wallets";
-import { toOutputScript } from "bitcoinjs-lib/src/address";
-import * as bitcoin from "bitcoinjs-lib";
-import * as ecc from "../../vendor/tiny-secp256k1/lib";
-
-bitcoin.initEccLib(ecc);
+import { PsbtInput, PsbtOutput } from "./psbt";
+import BigNumber from "bignumber.js";
+import {
+  braidDetailsToWalletConfig,
+  MultisigWalletConfig,
+} from "@caravan/wallets";
 
 export const idToHash = (txid: string): Buffer => {
   return Buffer.from(txid, "hex").reverse();
 };
 
-export interface PsbtInput {
-  hash: string | Buffer;
-  index: number;
-  transactionHex: string;
-  redeemScript?: Buffer;
-  witnessScript?: Buffer;
+// pulling functions from the old transactions/psbt file
+// so we can work with the fixtures and their deeply
+// nested objects
+export function psbtMultisigLock(multisig) {
+  if (!multisig) {
+    return {};
+  }
+  const multisigLock: any = {};
+
+  // eslint-disable-next-line default-case
+  switch (multisigAddressType(multisig)) {
+    case P2SH:
+      multisigLock.redeemScript = multisigRedeemScript(multisig).output;
+      break;
+    case P2WSH:
+      multisigLock.witnessScript = multisigWitnessScript(multisig).output;
+      break;
+    case P2SH_P2WSH: // need both
+      multisigLock.witnessScript = multisigWitnessScript(multisig).output;
+      multisigLock.redeemScript = multisigRedeemScript(multisig).output;
+      break;
+  }
+  return multisigLock;
+}
+
+export function getBip32Derivation(multisig, index = 0) {
+  if (!multisig) {
+    return;
+  }
+  // Already have one, return it
+  if (multisig.bip32Derivation) {
+    return multisig.bip32Derivation;
+  }
+  // Otherwise generate it
+  const config = JSON.parse(multisigBraidDetails(multisig));
+  const braid = generateBraid(
+    config.network,
+    config.addressType,
+    config.extendedPublicKeys,
+    config.requiredSigners,
+    config.index,
+  );
+  return generateBip32DerivationByIndex(braid, index);
+}
+
+export interface LegacyMultisig {
+  /**
+   * JSON stringified object with the following properties:
+   * braidDetails: {
+   *   network: Network;
+   *   addressType: number;
+   *   extendedPublicKeys: string[];
+   *   requiredSigners: number;
+   *   index: string;
+   * };
+   */
+  braidDetails: string;
   bip32Derivation?: {
-    masterFingerprint: Buffer;
+    masterFingerprint: string;
     path: string;
     pubkey: Buffer;
   }[];
 }
 
-export interface PsbtOutput {
+// This may be incomplete as the fixture objects are extremely opaque.
+// Hopefully we can build on this though as we work to convert more
+// code to fully typed Typescript.
+export interface LegacyInput {
+  txid: string;
+  index: number;
+  transactionHex: string;
+  amountSats: number | string;
+  multisig: LegacyMultisig;
+}
+
+export interface LegacyOutput {
   address: string;
-  value: number;
+  amountSats: number | string;
   bip32Derivation?: {
-    masterFingerprint: Buffer;
+    masterFingerprint: string;
     path: string;
     pubkey: Buffer;
   }[];
-  redeemScript?: Buffer;
   witnessScript?: Buffer;
+  redeemScript?: Buffer;
+  multisig?: LegacyMultisig;
 }
-/**
- * This function seeks to be an updated version of the legacy `unsignedMultisigPSBT` function
- * from @caravan/bitcoin.
- */
-export const getUnsignedMultisigPsbtV0 = ({
-  network,
-  inputs,
-  outputs,
-  multisigConfig,
-  includeGlobalXpubs = false,
-}: {
+
+export const convertLegacyInputs = (input: LegacyInput): PsbtInput => {
+  return {
+    hash: idToHash(input.txid),
+    index: input.index,
+    transactionHex: input.transactionHex,
+    bip32Derivation:
+      input.multisig.bip32Derivation || getBip32Derivation(input.multisig),
+    ...psbtMultisigLock(input.multisig),
+    spendingWallet: getWalletConfigFromInput(input),
+  };
+};
+
+export const convertLegacyOutputs = (output: LegacyOutput): PsbtOutput => {
+  return {
+    address: output.address,
+    value: new BigNumber(output.amountSats).toNumber(),
+    bip32Derivation:
+      output.bip32Derivation || getBip32Derivation(output.multisig),
+    witnessScript: output.witnessScript,
+    redeemScript: output.redeemScript,
+  };
+};
+
+export const getWalletConfigFromInput = (
+  input: LegacyInput,
+): MultisigWalletConfig => {
+  return braidDetailsToWalletConfig(JSON.parse(input.multisig.braidDetails));
+};
+
+export const psbtArgsFromFixture = (
+  fixture: any,
+): {
   network: Network;
   inputs: PsbtInput[];
   outputs: PsbtOutput[];
-  multisigConfig: MultisigWalletConfig;
-  includeGlobalXpubs?: boolean;
-}): Psbt => {
-  const psbt = new Psbt({ network: networkData(network) });
-  // should eventually support version 2, but to maintain compatibility with
-  // older api and existing fixtures, will keep with 1 for now
-  psbt.setVersion(1);
-  for (const input of inputs) {
-    const inputData = psbtInputFormatter(input, multisigConfig.addressType);
-    psbt.addInput(inputData);
-  }
-
-  const formatted = outputs.map((output) =>
-    psbtOutputFormatter(output, network),
-  );
-  psbt.addOutputs(formatted);
-  if (multisigConfig && includeGlobalXpubs) {
-    addGlobalXpubs(psbt, multisigConfig);
-  }
-
-  return psbt;
-};
-
-const psbtInputFormatter = (
-  input: PsbtInput,
-  addressType: MultisigAddressType,
-) => {
-  const tx = Transaction.fromHex(input.transactionHex);
-  const inputData: any = { ...input };
-  if (addressType === P2SH) {
-    const nonWitnessUtxo = tx.toBuffer();
-    inputData.nonWitnessUtxo = nonWitnessUtxo;
-  } else {
-    inputData.witnessUtxo = tx.outs[input.index];
-  }
-
-  // Delete key values with undefined values
-  Object.keys(inputData).forEach((key) => {
-    if (inputData[key] === undefined) {
-      delete inputData[key];
-    }
-  });
-
-  return inputData;
-};
-
-const psbtOutputFormatter = (output: PsbtOutput, network: Network) => {
-  const script = toOutputScript(output.address, networkData(network));
-  const outputData: any = {
-    ...output,
-    script,
-    value: output.value,
+} => {
+  return {
+    network: fixture.network as Network,
+    inputs: fixture.inputs.map(convertLegacyInputs),
+    outputs: fixture.outputs.map(convertLegacyOutputs),
   };
-  // Delete key values with undefined values
-  Object.keys(outputData).forEach((key) => {
-    if (outputData[key] === undefined) {
-      delete outputData[key];
-    }
-  });
-  return outputData;
-};
-
-export const addGlobalXpubs = (
-  psbt: Psbt,
-  multisigConfig: MultisigWalletConfig,
-) => {
-  const globalXpubs = multisigConfig.extendedPublicKeys.map(
-    ({ xfp, bip32Path, xpub }) => {
-      console.log("xpub:", xpub);
-      return {
-        extendedPubkey: ExtendedPublicKey.fromBase58(xpub).encode(),
-        masterFingerprint: xfp ? Buffer.from(xfp, "hex") : Buffer.alloc(0),
-        path: bip32Path || "",
-      };
-    },
-  );
-  psbt.updateGlobal({ globalXpub: globalXpubs });
 };
